Add Member command feature by extends from abstract addCommand
