+++ src/main/java/seedu/address/logic/commands/AddMemberCommand.java
@@ -1,0 +1,58 @@
+package seedu.address.logic.commands;
+
+import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
+import seedu.address.model.member.Member;
+
+import static java.util.Objects.requireNonNull;
+import static seedu.address.logic.parser.CliSyntax.*;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
+
+public class AddMemberCommand extends AddCommand{
+
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds a member to the ezFoodie. "
+            + "Parameters: "
+            + "-m "
+            + PREFIX_NAME + "NAME "
+            + PREFIX_PHONE + "PHONE "
+            + PREFIX_EMAIL + "EMAIL "
+            + PREFIX_ADDRESS + "ADDRESS "
+            + "[" + PREFIX_TAG + "TAG]...\n"
+            + "Example: " + COMMAND_WORD + " "
+            + PREFIX_NAME + "John Doe "
+            + PREFIX_PHONE + "98765432 "
+            + PREFIX_EMAIL + "johnd@example.com "
+            + PREFIX_ADDRESS + "311, Clementi Ave 2, #02-25 "
+            + PREFIX_TAG + "friends "
+            + PREFIX_TAG + "owesMoney";
+
+    public static final String MESSAGE_SUCCESS = "New member added: %1$s";
+    public static final String MESSAGE_DUPLICATE_MEMBER = "This member already exists in the ezFoodie";
+
+    private final Member toAdd;
+
+    public AddMemberCommand(Member member) {
+        requireNonNull(member);
+        toAdd = member;
+    }
+
+    @Override
+    public CommandResult execute(Model model) throws CommandException {
+        requireNonNull(model);
+
+        if (model.hasMember(toAdd)) {
+            throw new CommandException(MESSAGE_DUPLICATE_MEMBER);
+        }
+
+        model.addMember(toAdd);
+        return new CommandResult(String.format(MESSAGE_SUCCESS, toAdd));
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof AddCommand // instanceof handles nulls
+                && toAdd.equals(((AddMemberCommand) other).toAdd));
+    }
+
+}
