+++ src/test/java/seedu/address/logic/commands/AddCommandIntegrationTest.java
@@ -32,14 +32,14 @@ public class AddCommandIntegrationTest {
         Model expectedModel = new ModelManager(model.getEzFoodie(), new UserPrefs());
         expectedModel.addMember(validMember);
 
-        assertCommandSuccess(new AddCommand(validMember), model,
-                String.format(AddCommand.MESSAGE_SUCCESS, validMember), expectedModel);
+        assertCommandSuccess(new AddMemberCommand(validMember), model,
+                String.format(AddMemberCommand.MESSAGE_SUCCESS, validMember), expectedModel);
     }
 
     @Test
     public void execute_duplicateMember_throwsCommandException() {
         Member memberInList = model.getEzFoodie().getMemberList().get(0);
-        assertCommandFailure(new AddCommand(memberInList), model, AddCommand.MESSAGE_DUPLICATE_MEMBER);
+        assertCommandFailure(new AddMemberCommand(memberInList), model, AddMemberCommand.MESSAGE_DUPLICATE_MEMBER);
     }
 
 }
